Created BookServiceTest

+++ src/main/java/mate/project/store/service/impl/BookServiceImpl.java
@@ -60,6 +60,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteById(Long id) {
+        if (bookRepository.findById(id).isEmpty()) {
+            throw new EntityNotFoundException("Can't find a book with id " + id);
+        }
         bookRepository.deleteById(id);
     }
